# future: downgrade asserts to debug mode only