--- a/include/seastar/core/future.hh
+++ b/include/seastar/core/future.hh
@@ -413,7 +413,7 @@ struct future_state_base {
         any(state s) noexcept { st = s; }
         void set_exception(std::exception_ptr&& e) noexcept {
             new (&ex) std::exception_ptr(std::move(e));
-            SEASTAR_ASSERT(st >= state::exception_min);
+            assert(st >= state::exception_min);
         }
         any(std::exception_ptr&& e) noexcept {
             set_exception(std::move(e));
@@ -503,21 +503,21 @@ public:
     void ignore() noexcept;
 
     void set_exception(std::exception_ptr&& ex) noexcept {
-        SEASTAR_ASSERT(_u.st == state::future);
+        assert(_u.st == state::future);
         _u.set_exception(std::move(ex));
     }
     future_state_base& operator=(future_state_base&& x) noexcept = default;
     void set_exception(future_state_base&& state) noexcept {
-        SEASTAR_ASSERT(_u.st == state::future);
+        assert(_u.st == state::future);
         *this = std::move(state);
     }
     std::exception_ptr get_exception() && noexcept {
-        SEASTAR_ASSERT(_u.st >= state::exception_min);
+        assert(_u.st >= state::exception_min);
         // Move ex out so future::~future() knows we've handled it
         return _u.take_exception();
     }
     const std::exception_ptr& get_exception() const& noexcept {
-        SEASTAR_ASSERT(_u.st >= state::exception_min);
+        assert(_u.st >= state::exception_min);
         return _u.ex;
     }
     template <typename U>
@@ -609,7 +609,7 @@ struct future_state :  public future_state_base, private internal::uninitialized
     }
     template <typename... A>
     void set(A&&... a) noexcept {
-        SEASTAR_ASSERT(_u.st == state::future);
+        assert(_u.st == state::future);
         new (this) future_state(ready_future_marker(), std::forward<A>(a)...);
     }
     future_state(exception_future_marker, std::exception_ptr&& ex) noexcept : future_state_base(std::move(ex)) { }
@@ -618,21 +618,21 @@ struct future_state :  public future_state_base, private internal::uninitialized
     future_state(nested_exception_marker m, future_state_base&& old) noexcept : future_state_base(m, std::move(old)) { }
     future_state(nested_exception_marker m, future_state_base&& n, future_state_base&& old) noexcept : future_state_base(m, std::move(n), std::move(old)) { }
     T&& get_value() && noexcept {
-        SEASTAR_ASSERT(_u.st == state::result);
+        assert(_u.st == state::result);
         return static_cast<T&&>(this->uninitialized_get());
     }
     T&& take_value() && noexcept {
-        SEASTAR_ASSERT(_u.st == state::result);
+        assert(_u.st == state::result);
         _u.st = state::result_unavailable;
         return static_cast<T&&>(this->uninitialized_get());
     }
     template<typename U = T>
     const std::enable_if_t<std::is_copy_constructible_v<U>, U>& get_value() const& noexcept(copy_noexcept) {
-        SEASTAR_ASSERT(_u.st == state::result);
+        assert(_u.st == state::result);
         return this->uninitialized_get();
     }
     T&& take() && {
-        SEASTAR_ASSERT(available());
+        assert(available());
         if (_u.st >= state::exception_min) {
             std::move(*this).rethrow_exception();
         }
@@ -640,14 +640,14 @@ struct future_state :  public future_state_base, private internal::uninitialized
         return static_cast<T&&>(this->uninitialized_get());
     }
     T&& get() && {
-        SEASTAR_ASSERT(available());
+        assert(available());
         if (_u.st >= state::exception_min) {
             std::move(*this).rethrow_exception();
         }
         return static_cast<T&&>(this->uninitialized_get());
     }
     const T& get() const& {
-        SEASTAR_ASSERT(available());
+        assert(available());
         if (_u.st >= state::exception_min) {
             rethrow_exception();
         }
@@ -867,10 +867,7 @@ public:
         // The state can be null if the corresponding future has been
         // destroyed without producing a continuation.
         if (ptr) {
-            // FIXME: This is a fairly expensive assert. It would be a
-            // good candidate for being disabled in release builds if
-            // we had such an assert.
-            SEASTAR_ASSERT(ptr->_u.st == future_state_base::state::future);
+            assert(ptr->_u.st == future_state_base::state::future);
             new (ptr) future_state(std::move(state));
             make_ready<urgent::yes>();
         }
@@ -1222,7 +1219,7 @@ private:
             : _state(std::move(state)) {
     }
     internal::promise_base_with_type<T> get_promise() noexcept {
-        SEASTAR_ASSERT(!_promise);
+        assert(!_promise);
         return internal::promise_base_with_type<T>(this);
     }
     internal::promise_base_with_type<T>* detach_promise() noexcept {
@@ -1705,7 +1702,7 @@ public:
 
 private:
     void set_task(task& t) noexcept {
-        SEASTAR_ASSERT(_promise);
+        assert(_promise);
         _promise->set_task(&t);
     }
 
@@ -1714,7 +1711,7 @@ private:
             callback->set_state(get_available_state_ref());
             ::seastar::schedule(callback);
         } else {
-            SEASTAR_ASSERT(_promise);
+            assert(_promise);
             schedule(callback);
         }
 
@@ -1851,7 +1848,7 @@ template <typename T>
 inline
 future<T>
 promise<T>::get_future() noexcept {
-    SEASTAR_ASSERT(!this->_future && this->_state && !this->_task);
+    assert(!this->_future && this->_state && !this->_task);
     return future<T>(this);
 }