These assertions are very much on the hotpath, and in Redpanda, some of
our microbenchmarks had regressions from this change. Here's an example
microbenchmark: https://github.com/redpanda-data/redpanda/blob/dev/src/v/utils/tests/coro_bench.cc

Before this change:

```
[bench-wrapper] running benchmark : coro_rpbench_binary
[bench-wrapper] verbose=0, exec_in_shm=True, redirect_stderr=True
[bench-wrapper] rundir            : /dev/shm/vectorized_io
[bench-wrapper] command           : src/v/utils/tests/coro_rpbench_binary --blocked-reactor-notify-ms 2000000 --abort-on-seastar-bad-alloc -m1GiB -c1
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          1
random seed:              3258366462

test                               iterations      median         mad         min         max      allocs       tasks        inst      cycles
coro_bench.empty_cont               248220000     3.994ns     0.000ns     3.992ns     3.994ns       0.000       0.000        51.0         0.0
coro_bench.ss_now                   264140000     3.751ns     0.001ns     3.749ns     3.755ns       0.000       0.000        49.0         0.0
coro_bench.ss_now_collect           142070000     3.002ns     0.001ns     3.001ns     3.003ns       0.000       0.000        35.0         0.0
coro_bench.co_await_ready            52880000    18.841ns     0.013ns    18.818ns    18.860ns       1.000       0.000       256.1         0.0
coro_bench.co_await_ready_collect    39970000    17.685ns     0.008ns    17.673ns    17.693ns       1.000       0.000       231.7         0.0
coro_bench.nested_then5             118570000     8.395ns     0.000ns     8.395ns     8.396ns       0.000       0.000       106.1         0.0
coro_bench.chain_then5               35370000    28.231ns     0.003ns    28.227ns    28.241ns       0.000       0.000       249.2         0.0
coro_bench.nested_after_yield        14670000    68.380ns     0.012ns    68.368ns    68.660ns       2.000       3.000       882.3         0.0
coro_bench.chained_after_yield        3900000   255.790ns     0.744ns   254.805ns   256.733ns       9.000      10.000      3017.0         0.0
coro_bench.coro_after_yield          12610000    78.961ns     0.019ns    78.910ns    78.997ns       2.000       3.000       966.4         0.0
coro_bench.co_await_ready_nested3    19920000    49.956ns     0.005ns    49.951ns    49.990ns       3.000       0.000       690.3         0.0
coro_bench.co_await_ready_nested5    11180000    89.525ns     0.015ns    89.379ns    89.540ns       5.000       0.001      1125.5         0.0
coro_bench.empty_coro                60670000    16.437ns     0.009ns    16.427ns    16.448ns       1.000       0.000       222.1         0.0
coro_bench.never_awaits_collect      50950000    15.585ns     0.001ns    15.578ns    15.586ns       1.000       0.000       203.0         0.0
[bench-wrapper] benchmark completed successfully
```

After:

```
[bench-wrapper] running benchmark : coro_rpbench_binary
[bench-wrapper] verbose=0, exec_in_shm=True, redirect_stderr=True
[bench-wrapper] rundir            : /dev/shm/vectorized_io
[bench-wrapper] command           : src/v/utils/tests/coro_rpbench_binary --blocked-reactor-notify-ms 2000000 --abort-on-seastar-bad-alloc -m1GiB -c1
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          1
random seed:              3381407145

test                               iterations      median         mad         min         max      allocs       tasks        inst      cycles
coro_bench.empty_cont               246920000     4.011ns     0.004ns     4.004ns     4.021ns       0.000       0.000        48.0         0.0
coro_bench.ss_now                   281180000     3.522ns     0.000ns     3.520ns     3.528ns       0.000       0.000        46.0         0.0
coro_bench.ss_now_collect           157250000     2.528ns     0.000ns     2.527ns     2.529ns       0.000       0.000        35.0         0.0
coro_bench.co_await_ready            53230000    18.722ns     0.010ns    18.697ns    18.752ns       1.000       0.000       248.1         0.0
coro_bench.co_await_ready_collect    39990000    17.558ns     0.008ns    17.537ns    17.565ns       1.000       0.000       225.3         0.0
coro_bench.nested_then5             136750000     7.287ns     0.002ns     7.281ns     7.299ns       0.000       0.000        86.1         0.0
coro_bench.chain_then5               35040000    28.459ns     0.033ns    28.410ns    28.502ns       0.000       0.000       245.2         0.0
coro_bench.nested_after_yield        13350000    75.162ns     0.170ns    74.869ns    75.332ns       2.000       3.000       866.3         0.0
coro_bench.chained_after_yield        3740000   267.487ns     0.109ns   266.707ns   267.596ns       9.000      10.000      3017.1         0.0
coro_bench.coro_after_yield          10960000    91.208ns     0.017ns    91.191ns    91.255ns       2.000       3.000       935.4         0.0
coro_bench.co_await_ready_nested3    18720000    53.447ns     0.011ns    53.434ns    53.460ns       3.000       0.000       662.3         0.0
coro_bench.co_await_ready_nested5    11850000    84.850ns     0.018ns    84.794ns    84.871ns       5.000       0.001      1091.4         0.0
coro_bench.empty_coro                58740000    16.969ns     0.015ns    16.939ns    17.002ns       1.000       0.000       218.1         0.0
coro_bench.never_awaits_collect      50040000    15.770ns     0.001ns    15.768ns    15.777ns       1.000       0.000       202.0         0.0
[bench-wrapper] benchmark completed successfully
```

The after changes are pretty much the same as what they were before PR
#2646 (because Redpanda always compiled -DNDEBUG in release mode so the
asserts in Seastar were disabled).