@xemul 2025-04-29
Presumably, the higher the value in "iterations" column, the better. 
With that, here's the `after/before - 1.0` bar chart of the tests:
![image](https://github.com/user-attachments/assets/bcbc3377-28a9-48b0-8f80-6a5d3d5d60af)

There are 6 tests improved and 8 tests deteriorated, all changes are within 15% (not bad, indeed). 
However, I don't know how to measure if e.g. the 13% loss of coro_after_yield worth the 15% gain in nested_then5, but overall the changes doesn't look to be definitely worthwhile.