@avikivity 2025-04-29
> The after changes are pretty much the same as what they were before PR

I don't understand. If there's no change, what's the motivation? Or is it only in your internal testing and the before/after here are with assertions?