@avikivity 2025-04-29
I don't think those results are reliable, CPU frequency scaling is too random.

Better to repeat with the same random seed and with the cpu frequency fixed to some super low value (like 1GHz), or measure instructions (see linux_perf_event, perf_tests.hh, @travisdowns). I'm sympathetic to the change but would like to see better proof.