@michoecho 2025-05-16
> > The after changes are pretty much the same as what they were before PR
> 
> I don't understand. If there's no change, what's the motivation? Or is it only in your internal testing and the before/after here are with assertions?

OP meant that the results after this PR are as they were *before PR #2646*, not before *this* PR.