@travisdowns 2025-05-19
> Presumably, the higher the value in "iterations" column, the better.

"iterations" isn't a primary metric reflecting performance, it's just an outcome of internal auto-adjustment loop which tries to hit a specific runtime. It has some correlation with test runtime but it's better to use runtime (which confusingly are called "median", "max", etc), cycles or instructions. We usually use instructions because despite some problems they have very low noise and usually point in the right direction.